feat: migration to minor units

+++ AccountService/src/main/java/accountservice/dto/ReserveFundsForTransactionCommand.java
@@ -1,11 +1,10 @@
 package accountservice.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ReserveFundsForTransactionCommand(
     UUID sourceAccountId,
     UUID targetAccountId,
-    BigDecimal minorUnits,
+    Long minorUnits,
     UUID transactionId,
     UUID sourceAuthUserId) {}
