+++ AccountService/src/main/java/accountservice/dto/UpdateAccountBalanceCommand.java
@@ -1,8 +1,7 @@
 package accountservice.dto;
 
 import jakarta.validation.constraints.Positive;
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record UpdateAccountBalanceCommand(
-    UUID accountId, @Positive BigDecimal minorUnits, UUID authUserId) {}
+    UUID accountId, @Positive Long minorUnits, UUID authUserId) {}
