+++ ApiGateway/src/main/java/apigateway/dto/transaction/TransactionStatusResponseDto.java
@@ -2,10 +2,9 @@ package apigateway.dto.transaction;
 
 import enums.common.Currency;
 import enums.transaction.TransactionStatus;
-import java.math.BigDecimal;
 
 public record TransactionStatusResponseDto(
-    BigDecimal minorUnits,
+    Long minorUnits,
     Currency currency,
     TransactionStatus status,
     TransactionStatusAccountResponseDto sourceAccount,
