+++ CardService/src/main/java/cardservice/dto/UpdateCardCommand.java
@@ -1,13 +1,12 @@
 package cardservice.dto;
 
 import enums.card.CardStatus;
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record UpdateCardCommand(
     UUID cardId,
     CardStatus status,
-    BigDecimal dailyLimitMinorUnits,
-    BigDecimal monthlyLimitMinorUnits,
+    Long dailyLimitMinorUnits,
+    Long monthlyLimitMinorUnits,
     UUID authUserId,
     String role) {}
