+++ CardService/src/main/java/cardservice/entity/CardLimitHoldEntity.java
@@ -2,7 +2,6 @@ package cardservice.entity;
 
 import enums.account.ReservationStatus;
 import jakarta.persistence.*;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Getter;
@@ -25,7 +24,7 @@ public class CardLimitHoldEntity {
   private UUID transactionId;
 
   @Column(name = "minor_units", nullable = false)
-  private BigDecimal minorUnits;
+  private Long minorUnits;
 
   @Enumerated(EnumType.STRING)
   @Column(name = "status", nullable = false)
