+++ TransactionService/src/main/java/transactionservice/dto/CreateTransactionCommand.java
@@ -1,13 +1,12 @@
 package transactionservice.dto;
 
 import enums.common.Currency;
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateTransactionCommand(
     UUID sourceAccountId,
     UUID targetAccountId,
-    BigDecimal minorUnits,
+    Long minorUnits,
     Currency currency,
     UUID idempotencyKey,
     UUID sourceAuthUserId,
